+++ example/lib/src/common/constants.dart
@@ -1,0 +1,2 @@
+/// Open Weather Map API Key.
+const String openWeatherApiKey = '21450ca9db82aa862173bf64669fe6a6';