+++ example/lib/src/not_found/not_found_view.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class NotFoundView extends StatelessWidget {
+  const NotFoundView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text(
+          '404: Page not found',
+          style: TextStyle(fontSize: 25.0),
+        ),
+      ),
+    );
+  }
+}