+++ lib/open_weather_map_client.dart
@@ -1,8 +1,13 @@
 library open_weather_map_client;
 
 export 'src/common/constants.dart';
+export 'src/model/api_error.dart';
 export 'src/model/city.dart';
+export 'src/model/condition.dart';
+export 'src/model/coordinates.dart';
 export 'src/model/country.dart';
-export 'src/model/weather.dart';
+export 'src/model/current_weather.dart';
+export 'src/model/detail.dart';
+export 'src/model/rain.dart';
+export 'src/model/wind.dart';
 export 'src/open_weather_map.dart';
-export 'src/open_weather_map_provider.dart';